requirements.txt
# Core Libraries
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
pyyaml==6.0

# Signal Processing
pywavelets # For wavelet transforms
PyWavelets==1.4.1

# Machine Learning / Deep Learning
# tensorflow # or pytorch - uncomment based on choice
keras # If using TensorFlow backend
torch==2.0.1

# Hyperparameter Optimization
scikit-optimize # For Bayesian Optimization (or optuna, hyperopt)
optuna==3.3.0

# Other Utilities
tqdm # Progress bars
tqdm==4.65.0

# Add specific versions if needed, e.g., numpy==1.23.5
scipy==1.10.1 